requirements.txt
streamlit
replicate
python-dotenv
Pillow
requests
